feat: 새로운 구조로 Flask 앱 및 API 설정을 위한 초기화 파일 추가 - repository_api.py 파일 생성 - 전체적인 코드 구조 개선 및 모듈화 진행
