--- api/__init__.py
@@ -1,1 +1,0 @@
-# API 모듈 패키지 초기화 파일
