+++ app/core/__init__.py
@@ -1,0 +1,1 @@
+# This file makes the 'core' directory a Python package.
