+++ app/core/exceptions.py
@@ -42,4 +42,4 @@ class ServiceError(Exception):
 class ValidationError(ValueError):
     """입력 값 검증 오류를 위한 사용자 정의 예외 클래스"""
 
-    pass
+    pass
