+++ app/services/embeddings.py
@@ -8,10 +8,10 @@ from google import genai
 from google.genai import types
 from langchain_core.embeddings import Embeddings
 
-from config import Config
-from common.exceptions import EmbeddingError
+# 수정: config 및 예외 클래스 임포트 경로 변경
+from app.core.config import Config
+from app.core.exceptions import EmbeddingError
 
-# 로거 설정
 logger = logging.getLogger(__name__)
 
 
@@ -29,13 +29,27 @@ class GeminiAPIEmbeddings(Embeddings):
         self.query_task_type = query_task_type
         self.max_retries = Config.MAX_RETRIES
 
-        # 2개의 클라이언트 초기화 (각각 다른 API 키 사용)
-        self.client1 = genai.Client(api_key=Config.GEMINI_API_KEY1)
-        self.client2 = genai.Client(api_key=Config.GEMINI_API_KEY2)
-        self.clients = [self.client1, self.client2]
+        # API 키 로드 실패 시 앱 로딩 단계에서 알 수 있도록 try-except 추가
+        try:
+            self.client1 = genai.Client(api_key=Config.GEMINI_API_KEY1)
+            self.client2 = genai.Client(api_key=Config.GEMINI_API_KEY2)
+            self.clients = [self.client1, self.client2]
+            if not Config.GEMINI_API_KEY1 or not Config.GEMINI_API_KEY2:
+                logger.warning("하나 이상의 Gemini API 키가 설정되지 않았습니다. 기능이 제한될 수 있습니다.")
+        except Exception as e:
+            logger.error(f"Gemini 클라이언트 초기화 실패: {e}. API 키 설정을 확인하세요.")
+            # 두 클라이언트 모두 None으로 설정하거나, 에러를 발생시켜 앱 실행을 중단할 수 있습니다.
+            self.client1 = None
+            self.client2 = None
+            self.clients = [] # 빈 리스트로 설정하여 이후 로직에서 문제 발생 방지
+            # raise EmbeddingError("Gemini API 클라이언트 초기화 실패") # 필요시 앱 중단
 
-        # 스레드 안전성을 위한 락
         self.lock = Lock()
+
+    def _check_clients_available(self):
+        if not self.clients or not all(self.clients):
+            raise EmbeddingError("Gemini API 클라이언트가 제대로 초기화되지 않았습니다.")
+        return True
 
     def _calculate_sleep_time(self, is_quota_error):
         """오류 유형별 API 재시도 대기 시간 계산"""
@@ -54,6 +68,7 @@ class GeminiAPIEmbeddings(Embeddings):
         self, client, batch_texts, batch_start_idx, batch_num, total_batches
     ):
         """개별 클라이언트로 배치 처리"""
+        self._check_clients_available()
         retries_count = 0
 
         while retries_count < self.max_retries:
@@ -65,7 +80,6 @@ class GeminiAPIEmbeddings(Embeddings):
                         output_dimensionality=Config.EMBEDDING_DIMENSION
                     ),
                 )
-                # 성공한 임베딩과 인덱스 정보 반환
                 embeddings = [emb.values for emb in result.embeddings]
                 success_indices = list(
                     range(batch_start_idx, batch_start_idx + len(batch_texts))
@@ -73,13 +87,11 @@ class GeminiAPIEmbeddings(Embeddings):
 
                 with self.lock:
                     logger.debug(f"배치 {batch_num} 임베딩 성공 ({len(embeddings)}개)")
-                    # 마지막 배치가 아닌 경우에만 대기 메시지 출력
                     if batch_num < total_batches:
                         logger.info(
                             f"성공적인 임베딩 후 {Config.SUCCESS_SLEEP_TIME}초 대기합니다."
                         )
 
-                # 마지막 배치가 아닌 경우에만 대기
                 if batch_num < total_batches:
                     time.sleep(Config.SUCCESS_SLEEP_TIME)
 
@@ -106,8 +118,6 @@ class GeminiAPIEmbeddings(Embeddings):
                         logger.error(
                             f"배치 {batch_num} 임베딩 최종 실패 (최대 재시도 도달)"
                         )
-
-                    # 실패한 인덱스 정보 반환
                     failed_indices = list(
                         range(batch_start_idx, batch_start_idx + len(batch_texts))
                     )
@@ -117,7 +127,6 @@ class GeminiAPIEmbeddings(Embeddings):
                         "batch_num": batch_num,
                     }
 
-        # 이론상 도달 불가
         failed_indices = list(
             range(batch_start_idx, batch_start_idx + len(batch_texts))
         )
@@ -129,20 +138,27 @@ class GeminiAPIEmbeddings(Embeddings):
 
     def embed_documents(self, texts):
         """다수 문서 임베딩 (병렬 처리)"""
+        self._check_clients_available()
         if not texts:
             logger.info("빈 문서 리스트입니다.")
             return [], []
 
-        successful_embeddings = [None] * len(texts)  # 원본 순서 유지용
+        successful_embeddings = [None] * len(texts)
         failed_original_indices = []
 
         batch_size = Config.EMBEDDING_BATCH_SIZE
         total_batches = (len(texts) + batch_size - 1) // batch_size
 
-        logger.info(f"병렬 임베딩 시작. 총 배치: {total_batches}, 클라이언트: 2개")
+        logger.info(f"병렬 임베딩 시작. 총 배치: {total_batches}, 클라이언트: {len(self.clients)}개")
+        if not self.clients : # 클라이언트가 하나도 없으면 단일 스레드로 (혹은 오류)
+             logger.warning("사용 가능한 Gemini 클라이언트가 없어 단일 스레드로 임베딩을 시도합니다.")
+             # 이 경우 _process_batch_with_client를 직접 호출하거나, 다른 로직 필요
+             # 여기서는 간단히 빈 결과 반환 또는 예외 발생
+             if len(texts) > 0:
+                 raise EmbeddingError("Gemini 클라이언트 없이 문서 임베딩 불가")
+             return [], []
 
-        # 병렬 처리를 위한 작업 생성
-        with ThreadPoolExecutor(max_workers=2) as executor:
+        with ThreadPoolExecutor(max_workers=len(self.clients)) as executor:
             futures = []
 
             for batch_idx in range(0, len(texts), batch_size):
@@ -151,10 +167,8 @@ class GeminiAPIEmbeddings(Embeddings):
                     continue
 
                 current_batch_num = batch_idx // batch_size + 1
-                # 배치를 2개 클라이언트에 번갈아 할당
-                client = self.clients[current_batch_num % 2]
+                client = self.clients[current_batch_num % len(self.clients)]
 
-                # 병렬 작업 제출 (total_batches 전달)
                 future = executor.submit(
                     self._process_batch_with_client,
                     client,
@@ -165,29 +179,23 @@ class GeminiAPIEmbeddings(Embeddings):
                 )
                 futures.append(future)
 
-            # 결과 수집
             successful_count = 0
             failed_count = 0
 
             for future in as_completed(futures):
                 try:
                     result = future.result()
-
                     if result["success"]:
-                        # 성공한 임베딩을 원본 위치에 저장
                         for i, embedding in enumerate(result["embeddings"]):
                             original_idx = result["indices"][i]
                             successful_embeddings[original_idx] = embedding
                         successful_count += len(result["embeddings"])
                     else:
-                        # 실패한 인덱스 기록
                         failed_original_indices.extend(result["failed_indices"])
                         failed_count += len(result["failed_indices"])
-
                 except Exception as e:
                     logger.error(f"병렬 처리 중 예기치 않은 오류: {e}")
 
-        # None 값 제거 및 최종 결과 생성
         final_embeddings = [emb for emb in successful_embeddings if emb is not None]
 
         logger.info(
@@ -201,25 +209,27 @@ class GeminiAPIEmbeddings(Embeddings):
         return final_embeddings, sorted(list(set(failed_original_indices)))
 
     def embed_query(self, text):
-        """단일 쿼리 임베딩 생성 (첫 번째 클라이언트 사용)"""
+        """단일 쿼리 임베딩 생성 (첫 번째 사용 가능한 클라이언트 사용)"""
+        self._check_clients_available()
+        # 첫번째 클라이언트가 없으면 두번째, 둘다 없으면 에러
+        active_client = self.client1 if self.client1 else self.client2
+        if not active_client:
+            raise EmbeddingError("쿼리 임베딩을 위한 Gemini API 클라이언트가 없습니다.")
+            
         retries_count = 0
         logger.info(f"쿼리 임베딩 중 (Task: {self.query_task_type})...")
 
         while retries_count < self.max_retries:
             try:
-                result = self.client1.models.embed_content(
+                result = active_client.models.embed_content(
                     model=self.model_name,
                     contents=[text],
                     config=types.EmbedContentConfig(
                         output_dimensionality=Config.EMBEDDING_DIMENSION
                     ),
                 )
-
                 logger.info("쿼리 임베딩 성공.")
-                # 쿼리 임베딩 후에는 대기하지 않음
-
                 return result.embeddings[0].values
-
             except Exception as e_emb_query:
                 retries_count += 1
                 logger.warning(
@@ -237,5 +247,4 @@ class GeminiAPIEmbeddings(Embeddings):
                     raise EmbeddingError(
                         f"쿼리 임베딩에 실패했습니다 (재시도 {self.max_retries}회 시도 후): {text[:50]}..."
                     ) from e_emb_query
-
-        raise EmbeddingError("쿼리 임베딩 로직의 예기치 않은 종료.")
+        raise EmbeddingError("쿼리 임베딩 로직의 예기치 않은 종료.")
